[refactor[ separate service and interface
